(PC-14447) feat(BookingTickets): add barcode to BookingDetailsTicketContent
