+++ src/features/bookings/components/SwiperTickets/SwipperTicketPageTest.tsx
@@ -1,35 +1,16 @@
 import React from 'react'
 import styled from 'styled-components/native'
 
-import { bookingsSnapDouble } from 'features/bookings/api/bookingsSnapDouble'
-import { SwiperTickets } from 'features/bookings/components/SwiperTickets/SwiperTickets'
-import { TicketWithContent } from 'features/bookings/components/SwiperTickets/TicketWithContent'
+import { bookingsWithExternalBookingInformationsSnap } from 'features/bookings/api/bookingsSnapWithExternalBookingInformations'
+import { Ticket } from 'features/bookings/components/SwiperTickets/Ticket'
 
 const activationCodeFeatureEnabled = true
-const booking = bookingsSnapDouble.ongoing_bookings[0]
+const booking = bookingsWithExternalBookingInformationsSnap.ongoing_bookings[0]
 
 export function SwipperTicketPageTest() {
   return (
     <Container>
-      {booking.externalBookingsInfos ? (
-        booking.externalBookingsInfos.length === 1 ? (
-          <TicketWithContent
-            booking={booking}
-            activationCodeFeatureEnabled={activationCodeFeatureEnabled}
-            externalBookingsInfos={booking.externalBookingsInfos[0]}
-          />
-        ) : (
-          <SwiperTickets
-            booking={booking}
-            activationCodeFeatureEnabled={activationCodeFeatureEnabled}
-          />
-        )
-      ) : (
-        <TicketWithContent
-          booking={booking}
-          activationCodeFeatureEnabled={activationCodeFeatureEnabled}
-        />
-      )}
+      <Ticket booking={booking} activationCodeFeatureEnabled={activationCodeFeatureEnabled} />
     </Container>
   )
 }
