+++ src/features/bookings/components/SwiperTickets/getMultipleTickets.tsx
@@ -1,6 +1,6 @@
 import React from 'react'
 
-import { BookingReponseBis } from 'features/bookings/api/bookingsSnapDouble'
+import { BookingReponseBis } from 'features/bookings/api/bookingsSnapWithExternalBookingInformations'
 import { TicketWithContent } from 'features/bookings/components/SwiperTickets/TicketWithContent'
 
 export type TicketsProps = {
@@ -12,14 +12,22 @@ export function getMultipleTickets({ booking, activationCodeFeatureEnabled }: Ti
   return {
     tickets: [booking].map((ticket) => {
       if (ticket.externalBookingsInfos) {
-        return ticket.externalBookingsInfos.map((infos, index) => (
-          <TicketWithContent
-            key={index}
-            booking={ticket}
-            activationCodeFeatureEnabled={activationCodeFeatureEnabled}
-            externalBookingsInfos={infos}
-          />
-        ))
+        const totalSeatsIndex = ticket.externalBookingsInfos
+          .map((item) => item.seat)
+          .filter((seat) => seat).length
+        let currentSeatIndex = 0
+        return ticket.externalBookingsInfos.map((infos, index) => {
+          const seatIndex = infos.seat ? `${(currentSeatIndex += 1)}/${totalSeatsIndex}` : null
+          const externalBookingInfosWithCurrentSeatIndex = { ...infos, seatIndex }
+          return (
+            <TicketWithContent
+              key={index}
+              booking={ticket}
+              activationCodeFeatureEnabled={activationCodeFeatureEnabled}
+              externalBookingsInfos={externalBookingInfosWithCurrentSeatIndex}
+            />
+          )
+        })
       }
       return []
     }),
